% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_data.R
\name{manage_data}
\alias{manage_data}
\title{Manage Data}
\usage{
manage_data()
}
\description{
Check raw data, compare it to md5 sums, and download if necessary.
TODO: add in collapse of data folder to put  all english texts at top of data
TODO: fix paths. It's all screwed up due to conversion to a package.
}
