% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{next_word}
\alias{next_word}
\title{Predict Next Word}
\usage{
next_word(input_text)
}
\arguments{
\item{input_text}{a single element \code{\link{character}} vector with the text that you would
like to perform nextword prediction on.}
}
\value{
a \code{\link{character}} vector of suggested words in order

Uses ngram dictionary to compare the end of the provided string and provide suggestions.
}
\description{
Predict Next Word
}
