% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{NextWordR_server}
\alias{NextWordR_server}
\title{NextWordR Server Function Definition}
\usage{
NextWordR_server(input, output, session)
}
\description{
NextWordR Server Function Definition
}
\keyword{internal}
