% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{NextWordR_ui}
\alias{NextWordR_ui}
\title{NextWordR User Interface Definition}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
NextWordR_ui
}
\description{
NextWordR User Interface Definition
}
\keyword{internal}
