% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ngram_dictionary}
\alias{ngram_dictionary}
\title{ngram lookup dictionary}
\format{an object of class \code{\link{data.frame}}}
\source{
\href{https://d396qusza40orc.cloudfront.net/dsscapstone/dataset/Coursera-SwiftKey.zip}{
Processed from raw data made available by SwiftKey in cooperation with Coursera}
}
\usage{
ngram_dictionary
}
\description{
ngram lookup dictionary
}
\keyword{datasets}
