% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_data.R
\docType{data}
\name{raw_data_dir}
\alias{raw_data_dir}
\title{Raw data directory}
\format{An object of class \code{character} of length 1.}
\usage{
raw_data_dir
}
\description{
Where raw data is stored within this packages file structure
}
\keyword{internal}
