% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw_data_md5}
\alias{raw_data_md5}
\title{Raw Data MD5}
\format{a named \code{\link{character}} vector with md5 values and names of corresponding files}
\usage{
raw_data_md5
}
\description{
A set of named md5 hash strings which can be used to validate the contents of each of the raw
english language text files that are used to build the \code{\link{ngram_dictionary}}
}
\keyword{datasets}
