% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_data.R
\docType{data}
\name{raw_data_url}
\alias{raw_data_url}
\title{Raw data URL}
\format{An object of class \code{character} of length 1.}
\usage{
raw_data_url
}
\description{
Made available by SwiftKey to students in the Johns Hopkins Data Science Specialization on
Coursera.
}
\keyword{internal}
