% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{sugg_butt_listen}
\alias{sugg_butt_listen}
\title{Suggestion Button Observer Generator}
\usage{
sugg_butt_listen(input, output, session, suggestion)
}
\description{
Generate a listener for a single dynamically generated suggestion button
}
\keyword{internal}
