% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{sugg_butt_name}
\alias{sugg_butt_name}
\title{Suggestion Button Namer}
\usage{
sugg_butt_name(suggestion)
}
\description{
programmatic naming scheme to ensures consistent calls in button generation and listeners
}
\keyword{internal}
