% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_data.R
\name{manage_data}
\alias{manage_data}
\title{Manage Data}
\usage{
manage_data()
}
\description{
Check raw data, compare it to md5 sums, and download it if necessary. Management of the raw data
through functions helps ensure that the processed dictionary of ngrams that is built using this
data can be easily regenerated, without needed to pass around large intermediate data sets.
}
